Homogeneity index for a single image
The 'homogeneity_one_image' function computes the homogeneity index of a single image through the MASQH method. It allows users to chose an image from a folder and returns a single index between 0 and 1. The functions takes one input : 'exclude_zone'. This parameter should be 1 if zones need to be exluded and 0 if the image should be used without any zone excluded. Default value is 0.