Add files via upload
User manual for the MASQH standalone application : it describes the algorithm behind MASQH, details the installation procedures for MAC and WINDOWS and explains how to use the interface. If you need any help to use the software or if you have ideas to improve it, feel free to contact Fiona.milano@polymtl.ca.